feat: do not throw when encountering unknown block types
BREAKING CHANGE: When encountering unknown block types, the serializer will no
longer throw by default - instead if will render a hidden `div` with a message
noting that a serializer is missing. A message will also be logged to the
console. To use the old behavior of throwing un known types, pass
`ignoreUnknownTypes: false` as a property.